Get & Setter
Adding getter & setter & making a class public, to be able to get it
outside of the library.

The goal is to access HOTSPOT data to render them separately
